floor(x) --> int(floor(x))

+++ src/struphy/bsplines/bsplines_kernels.py
@@ -64,7 +64,7 @@ def find_span(t: 'float[:]', p: 'int', eta: 'float') -> 'int':
         returnVal = high - 1
     else:
         # Perform binary search
-        span = floor((low + high) / 2)
+        span = int(floor((low + high) / 2))
 
         while eta < t[span] or eta >= t[span + 1]:
 
@@ -72,7 +72,7 @@ def find_span(t: 'float[:]', p: 'int', eta: 'float') -> 'int':
                 high = span
             else:
                 low = span
-            span = floor((low + high) / 2)
+            span = int(floor((low + high) / 2))
 
         returnVal = span
